+++ R/MDP.R
@@ -25,14 +25,11 @@
 #' @param horizon numeric; Number of epochs. `Inf` specifies an infinite
 #' horizon.
 #' @param start Specifies in which state the MDP starts.
-#' @param max logical; is this a maximization problem (maximize reward) or a
-#' minimization (minimize cost specified in `reward`)?
 #' @param name a string to identify the MDP problem.
 #' @param x a `MDP` object.
 #' 
-#' @return The function returns an object of class POMDP which is list with an
-#' element called `'model'` containing a list with the model specification.
-#' [solve_POMDP()] reads the object and adds a list element called
+#' @return The function returns an object of class MDP which is list with 
+#'   the model specification. [solve_MDP()] reads the object and adds a list element called
 #' `'solution'`.
 #' @author Michael Hahsler
 #' @examples
@@ -51,7 +48,7 @@
 #'     "open-right" = "uniform",
 #'     "do-nothing" = "identity"),
 #'
-#'   # the rew helper expects: action, start.state, end.state, observation, value
+#'   # the reward helper R_() expects: action, start.state, end.state, observation, value
 #'   reward = rbind(
 #'     R_("open-left",  "tiger-left",  v = -100),
 #'     R_("open-left",  "tiger-right", v =   10),
@@ -63,22 +60,19 @@
 #'
 #' STiger
 #'
-#' STiger$model
-#'
-#' sol <- solve_MDP(STiger, e = 1e-7)
+#' sol <- solve_MDP(STiger, eps = 1e-7)
+#' sol
+#' 
 #' policy(sol)
+#' plot_value_function(sol)
 #' 
-#' # convert the MDP into a POMDP
+#' # convert the MDP into a POMDP and solve
 #' STiger_POMDP <- MDP2POMDP(STiger)
-#' STiger_POMDP
+#' sol2 <- solve_POMDP(STiger_POMDP)
+#' sol2 
 #' 
-#' # do 5 epochs with no discounting
-#' s <- solve_POMDP(STiger_POMDP, method = "enum", horizon = 5)
-#' s
-#'
-#' # value function and policy
-#' plot_value_function(s)
-#' policy(s)
+#' policy(sol2)
+#' plot_value_function(sol2)
 #' @export
 MDP <- function(states,
   actions,
@@ -87,12 +81,10 @@ MDP <- function(states,
   discount = .9,
   horizon = Inf,
   start = "uniform",
-  max = TRUE,
   name = NA) {
   
   ### unsolved pomdp model
   x <- list(
-    model = list(
       name = name,
       discount = discount,
       horizon = horizon,
@@ -100,12 +92,10 @@ MDP <- function(states,
       actions = actions,
       transition_prob = transition_prob,
       reward = reward,
-      start = start,
-      max = max
-    )
+      start = start
   )
   
-  class(x) <- "MDP"
+  class(x) <- list("MDP", "list")
   check_and_fix_MDP(x)
 }
 
@@ -117,12 +107,12 @@ print.MDP <- print.POMDP
 MDP2POMDP <- function(x) {
   # add an observation for each state and identity observation_probability for all actions ('*') 
   # (note: pomdp-solve does not support "identity" for observation_probs)
-  x$model$observations <- x$model$states
-  ident_matrix <- diag(length(x$model$states))
-  dimnames(ident_matrix) <- list(x$model$states, x$model$observations)
+  x$observations <- x$states
+  ident_matrix <- diag(length(x$states))
+  dimnames(ident_matrix) <- list(x$states, x$observations)
   
-  x$model$observation_prob <- list('*' = ident_matrix)
-  class(x) <- c("MDP", "POMDP")
+  x$observation_prob <- list('*' = ident_matrix)
+  class(x) <- c("MDP", "POMDP", "list")
   x
 }